SD4Mom: Exclusions
Exclusion for reason not delivery gets a different message. Exclusions take you back to the main menu. Can now load and edit a file with exclusions (ie, it only shows a warning if you are on the page with the exclusion questions). Answering the exclusion questions is now required. Fixed bug with interaction between required question and hiding a field with data in it. Set pilot field to 0 for all sites